+++ packages/core/realtime-js/src/RealtimeChannel.ts
@@ -518,7 +518,7 @@ export default class RealtimeChannel {
         event: 'track',
         payload,
       },
-      opts.timeout || this.timeout
+      opts
     )
   }